update: 코드 분리, 캘린더 모달 반응형 추가
